Change TechLooper Import User API to adapt to other datasources such as VNW
