+++ src/main/java/com/techlooper/entity/userimport/UserImportProfile.java
@@ -1,0 +1,7 @@
+package com.techlooper.entity.userimport;
+
+/**
+ * Created by NguyenDangKhoa on 2/11/15.
+ */
+public interface UserImportProfile {
+}
