+++ src/main/java/com/techlooper/service/UserImportDataProcessor.java
@@ -1,5 +1,6 @@
 package com.techlooper.service;
 
+import com.techlooper.entity.userimport.UserImportEntity;
 import com.techlooper.model.UserImportData;
 
 import java.util.List;
@@ -8,5 +9,5 @@ import java.util.List;
  * Created by NguyenDangKhoa on 1/30/15.
  */
 public interface UserImportDataProcessor {
-  void process(List<UserImportData> users);
+  List<UserImportEntity> process(List<UserImportData> users);
 }
